refactor: move v2 psi references tests to correct package
